Human Pose Estimation/requirements.txt
torch
torchvision
pillow
pycocotools
tqdm
opencv-python